Rename api provider exception and clean repo
