+++ src/exception/api_provider.py
@@ -1,0 +1,2 @@
+class ApiProviderException(Exception):
+    pass
